+++ bill/invoice_scenarios.go
@@ -23,6 +23,9 @@ func (inv *Invoice) ScenarioSummary() *tax.ScenarioSummary {
 }
 
 func (inv *Invoice) scenarioSummary(r *tax.Regime) *tax.ScenarioSummary {
+	if r == nil {
+		return nil
+	}
 	ss := r.ScenarioSet(ShortSchemaInvoice)
 	if ss == nil {
 		return nil
@@ -45,8 +48,7 @@ func (inv *Invoice) scenarioSummary(r *tax.Regime) *tax.ScenarioSummary {
 	return ss.SummaryFor(inv.Type, tags, exts)
 }
 
-func (inv *Invoice) prepareTags() error {
-	r := inv.TaxRegime()
+func (inv *Invoice) prepareTags(r *tax.Regime) error {
 	if r == nil {
 		return nil
 	}
@@ -54,7 +56,8 @@ func (inv *Invoice) prepareTags() error {
 		return nil
 	}
 
-	// First check the tags are all valid
+	// Check the tags are all valid and identified by the tax regime
+	// as acceptable for invoices.
 	for _, k := range inv.Tax.Tags {
 		if t := r.Tag(k); t == nil {
 			return validation.Errors{
@@ -64,15 +67,11 @@ func (inv *Invoice) prepareTags() error {
 			}
 		}
 	}
+
 	return nil
 }
 
-func (inv *Invoice) prepareScenarios() error {
-	r := inv.TaxRegime()
-	if r == nil {
-		return nil
-	}
-
+func (inv *Invoice) prepareScenarios(r *tax.Regime) error {
 	// Use the scenario summary to add any notes to the invoice
 	ss := inv.scenarioSummary(r)
 	if ss == nil {