+++ regimes/be/be.go
@@ -56,7 +56,7 @@ func Validate(doc interface{}) error {
 func Calculate(doc interface{}) error {
 	switch obj := doc.(type) {
 	case *tax.Identity:
-		return common.NormalizeTaxIdentity(obj)
+		return tax.NormalizeIdentity(obj)
 	}
 	return nil
 }