+++ regimes/nl/nl.go
@@ -113,7 +113,7 @@ func Validate(doc interface{}) error {
 func Calculate(doc interface{}) error {
 	switch obj := doc.(type) {
 	case *tax.Identity:
-		return common.NormalizeTaxIdentity(obj)
+		return tax.NormalizeIdentity(obj)
 	}
 	return nil
 }